Move methods from main to app.
